# docs/design/openapi.yaml
openapi: 3.1.0
info:
  title: RMS Implementation Architecture Baseline 2.0 — OpenAPI draft
  version: 2.0.0-draft
  description: Design only; NOT IMPLEMENTED. 业务来源唯一冻结Baseline1.3。服务器根/api/v1。全部保护操作检查session、启用状态、实际角色并集；自我评审等限制独立执行。请求只白名单，历史不可变，无generic
    status API。SQL BIGINT IDs/lockVersion采用十进制字符串。两个/auth技术GET为实现派生，不新增FR。数据库generated活动键不公开。密码原值与1..72 UTF-8 bytes技术政策已冻结；非admin assignee目录待决。
servers:
- url: /api/v1
security:
- cookieSession: []
...
paths:
...
  /requirements:
    get:
      operationId: listRequirements
      tags:
      - requirement
      summary: 分页组合查询，默认排除逻辑撤销需求，空items正常
      description: 分页组合查询，默认排除逻辑撤销需求，空items正常。仅启用账号；服务端重新检查角色并集及独立业务限制。
      x-module: requirement
      x-fr:
      - FR-SEARCH-001
      - FR-SEARCH-002
      - FR-SEARCH-003
      - FR-SEARCH-004
      - FR-SEARCH-005
      - FR-SEARCH-006
      x-br:
      - BR-AUTH-001
      - BR-AUTH-002
      - BR-AUTH-003
      - BR-AUTH-004
      - BR-LIFE-001
      - BR-REQ-002
      - BR-REQ-008
      - BR-REQ-012
      x-uc:
      - UC-06
      x-ac:
      - AC-09
      x-required-roles: *id001
      x-transaction: READ_ONLY
      responses:
        '200':
          description: 成功；具体业务含义见操作说明
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequirementResponsePage'
        '400':
          description: 字段/完整性错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证/凭证错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限/启用状态/自我评审/CSRF拒绝
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 目标不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 修订/状态/唯一/基准/图锁冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: 内部错误；业务写全回滚
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      parameters:
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/Size'
      - name: keyword
        in: query
        required: false
        schema:
          type: string
        description: 编号/标题/描述关键词
      - name: status
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/RequirementStatus'
        description: status
      - name: kind
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/RequirementKind'
        description: kind
      - name: level
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/RequirementLevel'
        description: level
      - name: priority
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/Priority'
        description: priority
      - name: tag
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/Id'
        description: 标签ID；多个筛选使用AND，不改变业务语义
    post:
      operationId: createRequirement
      tags:
      - requirement
      summary: 创建DRAFT，服务端生成稳定编号和creator
      description: 创建DRAFT，服务端生成稳定编号和creator。仅启用账号；服务端重新检查角色并集及独立业务限制。
      x-module: requirement
      x-fr:
      - FR-AUDIT-001
      - FR-REQ-001
      - FR-REQ-005
      - FR-REQ-006
      x-br:
      - BR-AUDIT-001
      - BR-AUDIT-002
      - BR-AUDIT-003
      - BR-AUDIT-004
      - BR-AUDIT-006
      - BR-AUTH-001
      - BR-AUTH-002
      - BR-AUTH-003
      - BR-AUTH-004
      - BR-REQ-001
      - BR-REQ-002
      - BR-REQ-003
      - BR-REQ-004
      - BR-REQ-006
      - BR-REQ-007
      - BR-REQ-009
      - BR-REQ-013
      x-uc:
      - UC-03
      - UC-S1
      x-ac:
      - AC-02
      - AC-10
      x-required-roles: &id003
      - REQUIREMENT_ENGINEER
      x-transaction: APPLICATION_REQUIRED
      responses:
        '201':
          description: 成功；具体业务含义见操作说明
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequirementResponse'
        '400':
          description: 字段/完整性错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证/凭证错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限/启用状态/自我评审/CSRF拒绝
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 目标不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 修订/状态/唯一/基准/图锁冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: 内部错误；业务写全回滚
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      parameters:
      - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRequirementRequest'
  /requirements/{id}:
    get:
      operationId: getRequirement
      tags:
      - requirement
      summary: 需求详情及只读历史入口，撤销需求保留可访问只读历史
      description: 需求详情及只读历史入口，撤销需求保留可访问只读历史。仅启用账号；服务端重新检查角色并集及独立业务限制。
      x-module: requirement
      x-fr:
      - FR-REQ-002
      - FR-SEARCH-007
      - FR-VER-005
      x-br:
      - BR-AUDIT-005
      - BR-AUTH-001
      - BR-AUTH-002
      - BR-AUTH-003
      - BR-AUTH-004
      - BR-REQ-012
      - BR-VER-001
      - BR-VER-005
      - BR-VER-008
      - BR-VER-009
      - BR-VER-011
      x-uc:
      - UC-03
      - UC-06
      - UC-08
      x-ac:
      - AC-02
      - AC-05
      - AC-09
      x-required-roles: *id001
      x-transaction: READ_ONLY
      responses:
        '200':
          description: 成功；具体业务含义见操作说明
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequirementResponse'
        '400':
          description: 字段/完整性错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证/凭证错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限/启用状态/自我评审/CSRF拒绝
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 目标不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 修订/状态/唯一/基准/图锁冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: 内部错误；业务写全回滚
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Id'
  /requirements/{id}/content:
    patch:
      operationId: editRequirementContent
      tags:
      - requirement
      summary: 只编辑DRAFT八项内容，形成基线后不能直接修改
      description: 只编辑DRAFT八项内容，形成基线后不能直接修改。仅启用账号；服务端重新检查角色并集及独立业务限制。
      x-module: requirement
      x-fr:
      - FR-AUDIT-001
      - FR-REQ-003
      - FR-REQ-005
      - FR-REQ-006
      x-br:
      - BR-AUDIT-001
      - BR-AUDIT-002
      - BR-AUDIT-003
      - BR-AUDIT-004
      - BR-AUDIT-006
      - BR-AUTH-001
      - BR-AUTH-002
      - BR-AUTH-003
      - BR-AUTH-004
      - BR-REQ-002
      - BR-REQ-004
      - BR-REQ-005
      - BR-REQ-006
      - BR-REQ-007
      - BR-REQ-013
      x-uc:
      - UC-03
      - UC-S1
      x-ac:
      - AC-02
      - AC-10
      x-required-roles: *id003
      x-transaction: APPLICATION_REQUIRED
      responses:
        '200':
          description: 成功；具体业务含义见操作说明
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequirementResponse'
        '400':
          description: 字段/完整性错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证/凭证错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限/启用状态/自我评审/CSRF拒绝
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 目标不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 修订/状态/唯一/基准/图锁冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: 内部错误；业务写全回滚
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Id'
      - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditRequirementContentRequest'
  /requirements/{id}/metadata:
    patch:
      operationId: editRequirementMetadata
      tags:
      - requirement
      summary: 独立维护负责人及标签，新负责人须启用，不产生正式版本
      description: 独立维护负责人及标签，新负责人须启用，不产生正式版本。仅启用账号；服务端重新检查角色并集及独立业务限制。
      x-module: requirement
      x-fr:
      - FR-AUDIT-001
      - FR-REQ-007
      - FR-REQ-008
      x-br:
      - BR-AUDIT-001
      - BR-AUDIT-002
      - BR-AUDIT-003
      - BR-AUDIT-004
      - BR-AUDIT-006
      - BR-AUTH-001
      - BR-AUTH-002
      - BR-AUTH-003
      - BR-AUTH-004
      - BR-REQ-008
      x-uc:
      - UC-03
      - UC-S1
      x-ac:
      - AC-02
      - AC-10
      x-required-roles: *id003
      x-transaction: APPLICATION_REQUIRED
      responses:
        '200':
          description: 成功；具体业务含义见操作说明
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequirementResponse'
        '400':
          description: 字段/完整性错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证/凭证错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限/启用状态/自我评审/CSRF拒绝
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 目标不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 修订/状态/唯一/基准/图锁冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: 内部错误；业务写全回滚
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Id'
      - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditMetadataRequest'
  /tags:
    get:
      operationId: listTags
      tags:
      - requirement
      summary: 读取可复用标签，名称筛选
      description: 读取可复用标签，名称筛选。仅启用账号；服务端重新检查角色并集及独立业务限制。
      x-module: requirement
      x-fr:
      - FR-REQ-007
      - FR-SEARCH-006
      x-br:
      - BR-AUDIT-001
      - BR-AUTH-001
      - BR-AUTH-002
      - BR-AUTH-003
      - BR-AUTH-004
      - BR-REQ-008
      - BR-REQ-012
      x-uc:
      - UC-03
      - UC-06
      x-ac:
      - AC-02
      - AC-09
      x-required-roles: *id001
      x-transaction: READ_ONLY
      responses:
        '200':
          description: 成功；具体业务含义见操作说明
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagResponsePage'
        '400':
          description: 字段/完整性错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证/凭证错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限/启用状态/自我评审/CSRF拒绝
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 目标不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 修订/状态/唯一/基准/图锁冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: 内部错误；业务写全回滚
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      parameters:
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/Size'
      - name: name
        in: query
        required: false
        schema:
          type: string
          maxLength: 50
        description: name
    post:
      operationId: createTag
      tags:
      - requirement
      summary: 创建规范化唯一标签
      description: 创建规范化唯一标签。仅启用账号；服务端重新检查角色并集及独立业务限制。
      x-module: requirement
      x-fr:
      - FR-AUDIT-001
      - FR-REQ-007
      x-br:
      - BR-AUDIT-001
      - BR-AUDIT-002
      - BR-AUDIT-003
      - BR-AUDIT-004
      - BR-AUDIT-006
      - BR-AUTH-001
      - BR-AUTH-002
      - BR-AUTH-003
      - BR-AUTH-004
      - BR-REQ-008
      x-uc:
      - UC-03
      - UC-S1
      x-ac:
      - AC-02
      - AC-10
      x-required-roles: *id003
      x-transaction: APPLICATION_REQUIRED
      responses:
        '201':
          description: 成功；具体业务含义见操作说明
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagResponse'
        '400':
          description: 字段/完整性错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证/凭证错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限/启用状态/自我评审/CSRF拒绝
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 目标不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 修订/状态/唯一/基准/图锁冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: 内部错误；业务写全回滚
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      parameters:
      - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTagRequest'
  /tags/{id}:
    patch:
      operationId: editTag
      tags:
      - requirement
      summary: 更新标签名称或说明并审计，无Tag修订号字段
      description: 更新标签名称或说明并审计，无Tag修订号字段。仅启用账号；服务端重新检查角色并集及独立业务限制。
      x-module: requirement
      x-fr:
      - FR-AUDIT-001
      - FR-REQ-007
      x-br:
      - BR-AUDIT-001
      - BR-AUDIT-002
      - BR-AUDIT-003
      - BR-AUDIT-004
      - BR-AUDIT-006
      - BR-AUTH-001
      - BR-AUTH-002
      - BR-AUTH-003
      - BR-AUTH-004
      - BR-REQ-008
      x-uc:
      - UC-03
      - UC-S1
      x-ac:
      - AC-02
      - AC-10
      x-required-roles: *id003
      x-transaction: APPLICATION_REQUIRED
      responses:
        '200':
          description: 成功；具体业务含义见操作说明
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagResponse'
        '400':
          description: 字段/完整性错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证/凭证错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限/启用状态/自我评审/CSRF拒绝
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 目标不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 修订/状态/唯一/基准/图锁冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: 内部错误；业务写全回滚
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Id'
      - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditTagRequest'
  /requirements/{id}/withdraw:
    post:
      operationId: withdrawRequirement
      tags:
      - requirement
      summary: 图锁内确认从未提交且无有效关联的DRAFT，逻辑撤销
      description: 图锁内确认从未提交且无有效关联的DRAFT，逻辑撤销。仅启用账号；服务端重新检查角色并集及独立业务限制。
      x-module: requirement
      x-fr:
      - FR-AUDIT-001
      - FR-REQ-004
      x-br:
      - BR-AUDIT-001
      - BR-AUDIT-002
      - BR-AUDIT-003
      - BR-AUDIT-004
      - BR-AUDIT-006
      - BR-AUTH-001
      - BR-AUTH-002
      - BR-AUTH-003
      - BR-AUTH-004
      - BR-REQ-010
      - BR-REQ-011
      - BR-REQ-012
      x-uc:
      - UC-03
      - UC-S1
      x-ac:
      - AC-10
      - AC-12
      x-required-roles: *id003
      x-transaction: APPLICATION_REQUIRED
      responses:
        '200':
          description: 成功；具体业务含义见操作说明
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequirementResponse'
        '400':
          description: 字段/完整性错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证/凭证错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限/启用状态/自我评审/CSRF拒绝
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 目标不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 修订/状态/唯一/基准/图锁冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: 内部错误；业务写全回滚
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Id'
      - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LockCommand'
...
  /requirements/{id}/reopen:
    post:
      operationId: reopenRequirement
      tags:
      - requirement
      summary: 仅REJECTED回到DRAFT，保留评审与firstSubmittedAt
      description: 仅REJECTED回到DRAFT，保留评审与firstSubmittedAt。仅启用账号；服务端重新检查角色并集及独立业务限制。
      x-module: requirement
      x-fr:
      - FR-AUDIT-001
      - FR-AUDIT-002
      - FR-LIFE-001
      - FR-LIFE-003
      - FR-LIFE-006
      x-br:
      - BR-AUDIT-001
      - BR-AUDIT-002
      - BR-AUDIT-003
      - BR-AUDIT-004
      - BR-AUDIT-006
      - BR-AUTH-001
      - BR-AUTH-002
      - BR-AUTH-003
      - BR-AUTH-004
      - BR-LIFE-001
      - BR-LIFE-002
      - BR-LIFE-004
      - BR-LIFE-006
      - BR-LIFE-009
      - BR-REQ-011
      x-uc:
      - UC-04
      - UC-S1
      x-ac:
      - AC-03
      - AC-10
      x-required-roles: *id003
      x-transaction: APPLICATION_REQUIRED
      responses:
        '200':
          description: 成功；具体业务含义见操作说明
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequirementResponse'
        '400':
          description: 字段/完整性错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证/凭证错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限/启用状态/自我评审/CSRF拒绝
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 目标不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 修订/状态/唯一/基准/图锁冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: 内部错误；业务写全回滚
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Id'
      - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LockCommand'
  /requirements/{id}/implement:
    post:
      operationId: implementRequirement
      tags:
      - requirement
      summary: 仅合法前态并确认当前正式版本的implement，说明写审计
      description: 仅合法前态并确认当前正式版本的implement，说明写审计。仅启用账号；服务端重新检查角色并集及独立业务限制。
      x-module: requirement
      x-fr:
      - FR-AUDIT-001
      - FR-AUDIT-002
      - FR-LIFE-001
      - FR-LIFE-003
      - FR-LIFE-004
      x-br:
      - BR-AUDIT-001
      - BR-AUDIT-002
      - BR-AUDIT-003
      - BR-AUDIT-004
      - BR-AUDIT-006
      - BR-AUTH-001
      - BR-AUTH-002
      - BR-AUTH-003
      - BR-AUTH-004
      - BR-LIFE-001
      - BR-LIFE-002
      - BR-LIFE-004
      - BR-LIFE-007
      - BR-LIFE-009
      x-uc:
      - UC-04
      - UC-S1
      x-ac:
      - AC-03
      - AC-10
      x-required-roles: &id004
      - PROJECT_MEMBER
      x-transaction: APPLICATION_REQUIRED
      responses:
        '200':
          description: 成功；具体业务含义见操作说明
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequirementResponse'
        '400':
          description: 字段/完整性错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证/凭证错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限/启用状态/自我评审/CSRF拒绝
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 目标不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 修订/状态/唯一/基准/图锁冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: 内部错误；业务写全回滚
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Id'
      - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VersionConfirmationRequest'
...
components:
  securitySchemes:
    cookieSession:
      type: apiKey
      in: cookie
      name: JSESSIONID
      description: HttpOnly Secure SameSite=Lax，同源session；所有unsafe请求另需X-CSRF-TOKEN，包括login/logout。
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  schemas:
...
    RequirementLevel:
      type: string
      enum:
      - BUSINESS
      - USER
      - SYSTEM
    RequirementKind:
      type: string
      enum:
      - FUNCTIONAL
      - QUALITY
      - CONSTRAINT
    Priority:
      type: string
      enum:
      - LOW
      - MEDIUM
      - HIGH
      - CRITICAL
    RequirementStatus:
      type: string
      enum:
      - DRAFT
      - UNDER_REVIEW
      - APPROVED
      - REJECTED
      - IMPLEMENTED
      - VERIFIED
...
    ApiError:
      type: object
      additionalProperties: false
      properties:
        code:
          type: string
          enum:
          - INVALID_INPUT
          - INCOMPLETE_CONTENT
          - UNAUTHENTICATED
          - INVALID_CREDENTIALS
          - FORBIDDEN
          - ACCOUNT_DISABLED
          - SELF_REVIEW
          - NOT_AUTHOR
          - NOT_FOUND
          - LOCK_VERSION_CONFLICT
          - STATE_CONFLICT
          - BASE_VERSION_STALE
          - DUPLICATE
          - ACTIVE_CHANGE_EXISTS
          - ACTIVE_REVIEW_EXISTS
          - DEPENDENCY_CYCLE
          - GRAPH_BUSY
          - INTERNAL_ERROR
        message:
          type: string
        correlationId:
          type: string
        currentLockVersion:
          $ref: '#/components/schemas/LockVersion'
      required:
      - code
      - message
      - correlationId
      description: 安全错误，无SQL/stack/密码；currentLockVersion仅有明确受权资源的修订冲突才提供。
...
    RequirementResponsePage:
      type: object
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/RequirementResponse'
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
          maximum: 100
        totalElements:
          type: integer
          minimum: 0
      required:
      - items
      - page
      - size
      - totalElements
...